+++ pomme-client/src/net/handler.rs
@@ -7,7 +7,7 @@ use super::NetworkEvent;
 use super::commands::{CommandTree, SharedCommandTree};
 use super::sender::PacketSender;
 use crate::entity::components::Position;
-use crate::ui::server_list::{MotdSpan, format_motd_spans};
+use crate::ui::text::{TextSpan, format_text_spans};
 
 pub fn handle_game_packet(
     packet: &ClientboundGamePacket,
@@ -196,13 +196,13 @@ pub fn handle_game_packet(
             }
         }
         ClientboundGamePacket::SystemChat(p) if !p.overlay => {
-            send_chat(event_tx, format_motd_spans(&p.content));
+            send_chat(event_tx, format_text_spans(&p.content));
         }
         ClientboundGamePacket::PlayerChat(p) => {
-            send_chat(event_tx, format_motd_spans(&p.message()));
+            send_chat(event_tx, format_text_spans(&p.message()));
         }
         ClientboundGamePacket::DisguisedChat(p) => {
-            send_chat(event_tx, format_motd_spans(&p.message));
+            send_chat(event_tx, format_text_spans(&p.message));
         }
         ClientboundGamePacket::BlockUpdate(p) => {
             let _ = event_tx.try_send(NetworkEvent::BlockUpdate {
@@ -498,7 +498,7 @@ pub fn handle_game_packet(
     }
 }
 
-fn send_chat(event_tx: &Sender<NetworkEvent>, spans: Vec<MotdSpan>) {
+fn send_chat(event_tx: &Sender<NetworkEvent>, spans: Vec<TextSpan>) {
     let text: String = spans.iter().map(|s| s.text.as_str()).collect();
     tracing::info!("Chat: {text}");
     let _ = event_tx.try_send(NetworkEvent::ChatMessage { spans });
